New graph operators
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@8425 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ msi.gama.core/src/msi/gaml/operators/Graphs.java
@@ -295,7 +295,7 @@ public class Graphs {
 	@operator(value = "beta_index")
 	@doc(value = "returns the beta index of the graph (Measures the level of connectivity in a graph and is expressed by the relationship between the number of links (e) over the number of nodes (v) : beta = e/v.", examples = { "beta_index(graphEpidemio)" }, see = {
 		"alpha_index","gamma_index","nb_cycles","connectivity_index" })
-	public static int betaIndex(final IGraph graph) {
+	public static double betaIndex(final IGraph graph) {
 		if ( graph == null ) { throw GamaRuntimeException
 			.error("In the beta_index operator, the graph should not be null!"); }
 		return graph.getEdges().size() / graph.getVertices().size();
@@ -304,7 +304,7 @@ public class Graphs {
 	@operator(value = "gamma_index")
 	@doc(value = "returns the gamma index of the graph (A measure of connectivity that considers the relationship between the number of observed links and the number of possible links: gamma = e/(3 * (v - 2)) - for planar graph.", examples = { "gamma_index(graphEpidemio)" }, see = {
 		"alpha_index","beta_index","nb_cycles","connectivity_index" })
-	public static int gammaIndex(final IGraph graph) {
+	public static double gammaIndex(final IGraph graph) {
 		if ( graph == null ) { throw GamaRuntimeException
 			.error("In the gamma_index operator, the graph should not be null!"); }
 		return graph.getEdges().size() / (2 * graph.getVertices().size() - 5);
@@ -313,7 +313,7 @@ public class Graphs {
 	@operator(value = "connectivity_index")
 	@doc(value = "retruns a simple connetivity index. This number is estimated through the number of nodes (v) and of sub-graphs (p) : IC = (v - p) /(v - 1).", examples = { "connectivity_index(graphEpidemio)" }, see = {
 		"alpha_index","beta_index","gamma_index","nb_cycles" })
-	public static int connectivityIndex(final IGraph graph) {
+	public static double connectivityIndex(final IGraph graph) {
 		if ( graph == null ) { throw GamaRuntimeException
 			.error("In the connectivity_index operator, the graph should not be null!"); }
 		int S = graph.getVertices().size();
@@ -338,7 +338,7 @@ public class Graphs {
 	@operator(value = "alpha_index")
 	@doc(value = "returns the alpha index of the graph (measure of connectivity which evaluates the number of cycles in a graph in comparison with the maximum number of cycles. The higher the alpha index, the more a network is connected. alpha = nb_cycles / (2*S-5) - planar graph)", examples = { "alpha_index(graphEpidemio)" }, see = {
 		"beta_index","gamma_index","nb_cycles","connectivity_index" })
-	public static int alphaIndex(final IGraph graph) {
+	public static double alphaIndex(final IGraph graph) {
 		if ( graph == null ) { throw GamaRuntimeException
 			.error("In the alpha_index operator, the graph should not be null!"); }
 		int S = graph.getVertices().size();
